Add files via upload
this is it, bat was the file that makes it work so don't worry about that.
if anti virus does not allow this file to run let it. you need to run the plain file first and then reset your computer for it to create the dll file. i made it that way so please do it